Move SenateXmlError to a dependency-free concord.errors leaf
concord.models.votes borrowed SenateXmlError from the senate_xml network
client purely to raise it in SenateVoteDetail.from_senate_xml — a layering
inversion (data layer depending on an HTTP client). Once #122 wired
senate_xml into the observability ledger, that edge closed an import cycle
(models -> votes -> senate_xml -> observability -> models), which #122
worked around with a lazy import plus a regression guard in
tests/test_import_cycles.py.

Define SenateXmlError in a new dependency-free concord/errors.py leaf that
both the client and the model import top-level. The models -> senate_xml
edge disappears structurally and the lazy import reverts to a clean
top-level one; the error keeps one identity shared by the client (network
failures) and the model (XML-parse failures).

Part of #123 (smell #1, the layering inversion). The façade cleanup
(smell #2) is a follow-up PR. Plan: docs/plans/untangle-models-senate-xml.md
